Fix wrangler dev failing on the proxy's non-function named exports
workerd treats every named export of a Worker's entry module as a named
entrypoint and requires it to be a function or ExportedHandler, so the
exported constants (FREE_MODEL, MAX_TOKENS, ...) made `wrangler dev`
abort with "Incorrect type for map entry 'FREE_MODEL'". Move all
constants and helpers into src/proxy.ts and reduce src/index.ts to just
the default handler that delegates to handleFetch; the unit tests import
the helpers from src/proxy instead. Verified with `wrangler deploy
--dry-run`.
